Reorganize resource utils
