+++ src/utils/components/PolicyForm/PolicyWizard/InterfacesStep.tsx
@@ -10,7 +10,7 @@ import {
   V1NodeNetworkConfigurationPolicy,
 } from '@types';
 import { useNMStateTranslation } from '@utils/hooks/useNMStateTranslation';
-import { getPolicyInterfaces } from '@utils/policies/utils';
+import { getPolicyInterfaces } from '@utils/resources/policies/utils';
 
 import PolicyFormOVSBridgeMapping from '../components/PolicyFormOVSBridgeMapping';
 import { doesOVSBridgeExist } from '../utils';
