+++ src/utils/components/PolicyForm/components/PolicyFormOVSBridgeMapping.tsx
@@ -5,7 +5,7 @@ import { Button, ButtonVariant, Content, Popover, Title } from '@patternfly/reac
 import { HelpIcon, PlusCircleIcon } from '@patternfly/react-icons';
 import { V1NodeNetworkConfigurationPolicy } from '@types';
 import { useNMStateTranslation } from '@utils/hooks/useNMStateTranslation';
-import { OVN_BRIDGE_MAPPINGS } from '@utils/ovn/constants';
+import { OVN_BRIDGE_MAPPINGS } from '@utils/resources/ovn/constants';
 
 import PolicyFormOVSBridgeMappingExpandable from './PolicyFormOVSBridgeMappingExpandable';
