+++ src/utils/resources/interfaces/getters.ts
