+++ src/utils/resources/nns/getters.tsx
